[clang] allow global initializers to use non-constant static storage variables
This patch enhances the compatibility with GCC in Compound literals.

Add a new BoolFOption that is true by default, namely
allow-non-const-global-init.
If this option is set to true, the check for 6.5.2.5p3 while compiling
CPP files will be ignored, which is the default behavior of GCC, therefore
allowing global initializers to use non-constant static storage variables.
This is commonly harmless for single TU cases but might cause the static
initialization order fiasco if variables in other source files are used.

The test command in SemaCXX/compound-literal.cpp is changed, to ensure the
original test can be passed. Also, add two test files for setting the option
to both true and false.